rename context files to their function name & add styles folder
