+++ src/components/app.js
@@ -1,11 +1,9 @@
 import React, { useState, useEffect } from "react";
 
 import Header from "./layouts/header";
-import AddUser from "./users/addUser";
-import Modal from "./layouts/modal";
-import Users from "./users/users";
-import UsersContext from "../context/users";
-import AuthContext from "../context/auth";
+import UsersContext from "../context/usersContext";
+import AuthContext from "../context/authContext";
+import Main from "./layouts/main";
 
 function App() {
   const [auth, setAuth] = useState(false);
@@ -37,16 +35,7 @@ function App() {
           }}
         >
           <Header />
-          <div className="container mx-auto sm:px-20 px-6 font-IRANSans">
-            {/* Button for open modal add user */}
-            <AddUser />
-
-            {/* open modal (user form) */}
-            <Modal />
-
-            {/* users table (show users) */}
-            <Users />
-          </div>
+          <Main />
         </UsersContext.Provider>
       </AuthContext.Provider>
     </>
