+++ src/components/layouts/header.jsx
@@ -1,5 +1,5 @@
 import React, { useContext } from "react";
-import AuthContext from "../../context/auth";
+import AuthContext from "../../context/authContext";
 
 function Header() {
   const authContext = useContext(AuthContext);
