+++ src/components/styles/index.css
@@ -1,4 +1,4 @@
-@import url("./fonts//font.css");
+@import url("./../../fonts/font.css");
 
 @tailwind base;
 @tailwind components;
