+++ src/index.js
@@ -1,7 +1,7 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
 import App from "./components/app";
-import "./index.css";
+import "./components/styles/index.css";
 
 const root = ReactDOM.createRoot(document.getElementById("root"));
 root.render(<App />);
